Move some animation stuff around...
