+++ Assets/Script/Venue/VenueCamera/PosterizeComponent.cs
@@ -10,7 +10,7 @@ namespace YARG.Venue.VenueCamera
     [VolumeComponentMenuForRenderPipeline("Venue/Posterize", typeof(UniversalRenderPipeline))]
     public class PosterizeComponent : VolumeComponent, IPostProcessComponent
     {
-        public ClampedIntParameter Steps = new(0, 0, 20);
+        public ClampedIntParameter Steps = new(0, 0, 100);
 
         public bool IsActive() => Steps.overrideState && Steps.value > 0;
         public bool IsTileCompatible() => true;
